% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display.R
\name{data_ms2_by_plot}
\alias{data_ms2_by_plot}
\title{Plot MS2 spectrum with b, y ions.}
\usage{
data_ms2_by_plot(mzXML, scan_num, pep_seq = "", threshold = 0.5)
}
\arguments{
\item{mzXML}{A data set from data_read_mzXML function.}

\item{scan_num}{A integer.}

\item{pep_seq}{A string of peptide sequence.}

\item{threshold}{A integer of mz error with defalut .5.}
}
\value{
A list of ggplot2 object.
}
\description{
Plot MS2 spectrum with b, y ions.
}
\details{
Input a mzXML from data_read_mzXML function and scan number and peptide sequence, then return a list of ggplot objects.
}
