% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_display.R
\name{data_read_mzXML}
\alias{data_read_mzXML}
\title{Read MS data in mzXML format.}
\usage{
data_read_mzXML(filepath)
}
\arguments{
\item{filepath}{A string of mzXML file location.}
}
\value{
A list.
}
\description{
Read MS data in mzXML format.
}
\details{
Input a file path of mzXML format, then return a data set. The mzXML format file can be transformed wi MSConvert software.
}
